requirements.txt
streamlit
pymupdf
pytesseract
opencv-python-headless
pdf2image
pillow
requests